ISSUE-4: Fix bug in Private_GetCallerDirectory function
Private_GetCallerDirectory was not walking the call stack correctly.  As
a result it was returning the directory of the Pslogg module when a
Pslogg function was being called directly from the PowerShell console.

Now if a Pslogg function is called directly from the PowerShell
console, Private_GetCallerDirectory will return the current working
directory set in the PowerShell console.